WFO/django-staging/requirements.txt
APScheduler==3.7.0
boto3==1.12.19
Django==3.0.4
# psycopg2==2.8.4
psycopg2-binary==2.8.5
sqlparse==0.3.1
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
djangorestframework-jwt
django-rest-multiple-models
django-prometheus==2.0.0
requests
django-cors-middleware
stomp.py==7.0.0
aiostomp==1.6.2
django-redis==4.12.1
python-decouple==3.3
gunicorn==20.0.4
redis==3.0.1
redis-py-cluster==2.0.0
django-queryset-csv==1.1.0
python-dateutil==2.8.1
pandas==1.0.1
simplejson==3.16.0
mysqlclient==2.0.3
openpyxl==3.0.7
xlsxwriter==3.0.1
django-apscheduler==0.6.0
drf-yasg==1.17.1
coreapi==2.3.3
coreschema==0.0.4
python-ldap==3.0.0
eventlet==0.26.1
numpy==1.23.5
mssql-django==1.2
pyodbc==4.0.34